requirements.txt
anthropic>=0.40.0
playwright>=1.48.0
pdfminer.six>=20221105
python-docx>=1.1.0
google-api-python-client>=2.140.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
pyyaml>=6.0.1
rich>=13.7.0
click>=8.1.7
aiohttp>=3.10.0
beautifulsoup4>=4.12.0
lxml>=5.2.0
python-dotenv>=1.0.0
tenacity>=9.0.0
fake-useragent>=1.5.0
